refactor: extract shared date utils and action item types, add duration/recurrence to ScheduleInfoCard
- Extract getRelativeLabel, formatDateJa, formatTime, formatScheduleDateTime,
  formatDateShort, formatCommentDate into src/utils/date.ts and re-export from
  feature-specific utils files to eliminate duplication
- Unify getRelativeLabel behavior to dashboard version with tomorrow/yesterday labels
- Extract PendingActionItem and AllPendingActionItem types into src/types/action-item.ts,
  matching backend PendingActionItemSchema (organizer_id) and AllPendingActionItemSchema
  (counterpart_id) respectively
- Add optional duration_minutes and recurrence fields to ScheduleDetail type with
  TODO comment noting backend schema needs update
- Display duration and recurrence in ScheduleInfoCard
